pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "distill-feed"
version = "0.1.0"
description = "Agent-friendly CLI for feed and URL article digests"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Gary Chang" }]
dependencies = [
  "click>=8.1",
  "httpx>=0.27",
  "feedparser>=6.0",
  "trafilatura>=1.12",
  "openai>=1.60",
  "pydantic>=2.9",
  "pydantic-settings>=2.5",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "respx>=0.22",
  "ruff>=0.8",
]

[project.scripts]
distill-feed = "distill_feed.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
